+++ source/setup/containers/SaveCredentialsPage.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import VError from "verror";
+import { Layerr } from "layerr";
 import { Group } from "../../shared/library/buttercup.js";
 import SaveCredentialsPage from "../components/SaveCredentialsPage.js";
 import { getArchives } from "../../shared/selectors/archives.js";
@@ -67,7 +67,7 @@ export default connect(
                         })
                         .catch(err => {
                             dispatch(unsetBusy());
-                            const { authFailure = false } = VError.info(err);
+                            const { authFailure = false } = Layerr.info(err);
                             if (authFailure) {
                                 notifyWarning(
                                     "Authorisation failed",