+++ source/setup/containers/UnlockAllArchivesPage.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import VError from "verror";
+import { Layerr } from "layerr";
 import UnlockAllArchivesPage from "../components/UnlockAllArchivesPage.js";
 import { getArchives } from "../../shared/selectors/archives.js";
 import { unlockArchive } from "../library/messaging.js";
@@ -42,7 +42,7 @@ export default connect(
                 })
                 .catch(err => {
                     console.error(err);
-                    const { hush } = VError.info(err);
+                    const { hush } = Layerr.info(err);
                     if (hush) {
                         notifyWarning("Authorisation failed", "The credentials were invalid - re-authenticating");
                     } else {