+++ source/shared/library/config.js
@@ -1,5 +1,6 @@
 export const INITIAL_CONFIG = {
     darkMode: false,
+    dynamicIcons: "",
     showSaveDialog: "always", // always/unlocked/never
     autoUnlockVaults: true,
     autoLockVaults: "off"